In progress

+++ src/main/java/com/example/BillingCentreTest/controllers/TaskController.java
@@ -3,11 +3,13 @@ package com.example.BillingCentreTest.controllers;
 import com.example.BillingCentreTest.models.Task;
 import com.example.BillingCentreTest.services.TaskService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @RestController
@@ -25,32 +27,44 @@ public class TaskController {
     public List<Task> findAll() {
         return taskService.findAll();
     }
+
     @GetMapping("/byDate")
-    public List<Task> findbyDate(@PathVariable("date") Date date) {
+    public List<Task> findbyDate(@RequestParam("date") Date date) {
         return taskService.findbyDate(date);
     }
 
-    @GetMapping("/{id}")
-    public Task findOne(@PathVariable("id") UUID id) {
+    @GetMapping("/one")
+    public Task findOne(@RequestParam("id") long id) {
         return taskService.findOne(id);
     }
 
+//    @GetMapping()
+//    public List<String> findType() {
+//        return taskService.findType();
+//    }
 
     @PostMapping("/add")
-    public void create(@RequestBody Task task) {
+    public ResponseEntity<Task> create(@RequestBody Task task) {
         taskService.save(task);
-
+        return ResponseEntity.ok().build();
     }
 
 
-    @PatchMapping("/{id}")
-    public void update(@PathVariable("id") UUID id) {
-        Task task = taskService.findOne(id);
-        taskService.update(id, task);
+    @PatchMapping()
+    public ResponseEntity<Task> update(@RequestBody Task task) {
+        if (taskService.findOne(task.getId()) == null) {
+            return ResponseEntity.notFound().build();
+        }
+        taskService.save(task);
+        return ResponseEntity.ok().build();
     }
 
-    @DeleteMapping("/{id}")
-    public void delete(@PathVariable("id") UUID id) {
+    @DeleteMapping()
+    public ResponseEntity<Task> delete(@RequestParam("id") long id) {
+        if (taskService.findOne(id) == null) {
+            return ResponseEntity.notFound().build();
+        }
         taskService.delete(id);
+        return ResponseEntity.ok().build();
     }
 }
